requirements.txt
gluonnlp==0.8.3
mxnet-cu101==1.5.0
numpy==1.16.2
regex
tqdm
mxboard